requirements.txt
requests==2.22.0
djangorestframework_jwt==1.11.0
django_rest_swagger==2.2.0
Jinja2==2.10.1
SQLAlchemy==1.2.8
chardet==3.0.4
django_filter==2.1.0
paramiko==2.4.1
pandas==0.23.4
PyMySQL==0.9.2
django_redis==4.9.0
kubernetes==9.0.0
django_auth_ldap==1.7.0
psycopg2==2.7.3.1
Django==3.0
DBUtils==1.3
aliyun_python_sdk_cms==7.0.9
aliyun_python_sdk_core==2.13.14
aliyun_python_sdk_ecs==4.18.3
channels==2.4.0
django_filters==0.2.1
python_ldap==3.2.0
djangorestframework==3.11.0
aliyun-python-sdk-alidns==2.0.16
django-cors-headers==3.5.0
django-filter==2.1.0
jinja2==2.10.1
fabric2==2.5.0
cryptography==2.4.2
celery==4.3.1
django-celery-beat==1.1.1
django-celery-results==1.2.1
asgiref==3.3.1
django-auditlog==1.0a1
django-utils-six==2.0
python-gitlab==2.5.0
djangorestframework-jwt==1.11.0
